+++ src/components/Button/Button.types.ts
@@ -58,6 +58,13 @@ export type ButtonProps = StyleOverrideProps & {
    */
   disabled?: boolean;
   /**
+   * Whether the button should take up the full width of its container.
+   *
+   * @default false
+   */
+  fullWidth?: boolean;
+
+  /**
    * Whether to display a loading spinner or not.
    *
    * Setting this to `true` will also prevent any `onClick` events from being triggered.